blackbox: Make writes async-signal safe
Use a thread_local variable to record a lock acquisition for the
blackbox write mutex. This can then be used to avoid deadlocks when
performing write_locked calls from a signal handler, as we can detect
presence on a critical section when interrupting it from a signal
handler (or nested signal handlers with interrupted handler performing a
write_locked operation).

This approach also improves the dumping operation, which can rely on
this thread_local variable to determine whether it can safely acquire
the mutex or not, instead of performing a try lock operation which can
spuriously fail due to contention.

To enable this and return a sane error, write locked is now changed to
always return an int, instead of the return type of the callable
template parameter. Since it's only used from the insert function, this
seems reasonable, but we still allow a wider set of return types by
relying on the signed_integral concept. If the return value is not a
signed integral, compilation will fail.

Note that we need to mark lock_taken even in dump, for two reasons:
First, write_locked may interrupt dump, just as dump may interrupt
write_locked. Second, we may begin waiting in write_locked after
interrupting dump while the owner (lower context) may never make
progress because we keep waiting while having interrupted it.

To keep using RAII pattern in dump, create a mini SignalSafeLock class
with a factory method to return an optional containing the class such
that lock_taken is set and reset before and after the lock acquisition
and release. For consistency, switch write_locked over to use this new
class as well.

Also use EDEADLK rather than EWOULDBLOCK in both cases (write_locked and
dump) for a more meaningful string error.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>